requirements.txt
opencv-python
numpy
matplotlib
tensorflow
flask
ffmpeg-python
scikit-video
pandas
scikit-learn
